+++ CHANGELOG.md
@@ -1,7 +1,7 @@
-### Unreleased
+### v2.8.0
 * bootstrap: remove unnecessary subscription to isp-gate
-* csvwriter : added new option (`WithGzipCompression`)
-* streaming : added new writer for streaming
+* csvwriter: added new option (`WithGzipCompression`)
+* streaming: added new writer for streaming
 * added package `scripts` with goja wrapping
 ### v2.7.0
 * bootstrap: add handling errors from `ackEvent` which resolved deadlock of **main select**